Move the home accordions to a separate view
